refactor(cl): convert PeopleTestCase to a mixin
