+++ cl/lib/test_helpers.py
@@ -15,18 +15,12 @@ from cl.audio.factories import AudioFactory
 from cl.audio.models import Audio
 from cl.lib.utils import deepgetattr
 from cl.people_db.factories import (
-    ABARatingFactory,
     AttorneyFactory,
     AttorneyOrganizationFactory,
-    EducationFactory,
     PartyFactory,
     PartyTypeFactory,
     PersonFactory,
-    PoliticalAffiliationFactory,
-    PositionFactory,
-    SchoolFactory,
 )
-from cl.people_db.models import Race
 from cl.search.constants import o_type_index_map
 from cl.search.docket_sources import DocketSources
 from cl.search.documents import DocketDocument
@@ -789,152 +783,6 @@ audio_v4_fields.update(
         "meta": [],  # type: ignore
     }
 )
-
-
-class PeopleTestCase(TestCase):
-    """People test case factories"""
-
-    @classmethod
-    def setUpTestData(cls):
-        cls.w_race, _ = Race.objects.get_or_create(race="w")
-        cls.b_race, _ = Race.objects.get_or_create(race="b")
-        cls.person_1 = PersonFactory.create(
-            gender="m",
-            name_first="Bill",
-            name_last="Clinton",
-        )
-        cls.person_1.race.add(cls.w_race)
-
-        cls.person_2 = PersonFactory.create(
-            gender="f",
-            name_first="Judith",
-            name_last="Sheindlin",
-            name_suffix="2",
-            date_dob=datetime.date(1942, 10, 21),
-            date_dod=datetime.date(2020, 11, 25),
-            date_granularity_dob="%Y-%m-%d",
-            date_granularity_dod="%Y-%m-%d",
-            name_middle="Susan",
-            dob_city="Brookyln",
-            dob_state="NY",
-            fjc_id=19832,
-        )
-        cls.person_2.race.add(cls.w_race)
-        cls.person_2.race.add(cls.b_race)
-
-        cls.person_3 = PersonFactory.create(
-            gender="f",
-            name_first="Sheindlin",
-            name_last="Judith",
-            date_dob=datetime.date(1945, 11, 20),
-            date_granularity_dob="%Y-%m-%d",
-            name_middle="Olivia",
-            dob_city="Queens",
-            dob_state="NY",
-        )
-        cls.person_3.race.add(cls.w_race)
-
-        cls.position_1 = PositionFactory.create(
-            date_granularity_start="%Y-%m-%d",
-            date_start=datetime.date(1993, 1, 20),
-            date_retirement=datetime.date(2001, 1, 20),
-            termination_reason="retire_mand",
-            position_type="pres",
-            person=cls.person_1,
-            how_selected="e_part",
-        )
-        cls.position_2 = PositionFactory.create(
-            date_granularity_start="%Y-%m-%d",
-            court=cls.court_1,
-            date_start=datetime.date(2015, 12, 14),
-            predecessor=cls.person_2,
-            appointer=cls.position_1,
-            judicial_committee_action="no_rep",
-            termination_reason="retire_mand",
-            position_type="c-jud",
-            person=cls.person_2,
-            how_selected="e_part",
-            nomination_process="fed_senate",
-            date_elected=datetime.date(2015, 11, 12),
-            date_confirmation=datetime.date(2015, 11, 14),
-            date_termination=datetime.date(2018, 10, 14),
-            date_granularity_termination="%Y-%m-%d",
-            date_hearing=datetime.date(2021, 10, 14),
-            date_judicial_committee_action=datetime.date(2022, 10, 14),
-            date_recess_appointment=datetime.date(2013, 10, 14),
-            date_referred_to_judicial_committee=datetime.date(2010, 10, 14),
-            date_retirement=datetime.date(2023, 10, 14),
-        )
-        cls.position_3 = PositionFactory.create(
-            date_granularity_start="%Y-%m-%d",
-            date_start=datetime.date(2015, 12, 14),
-            organization_name="Pants, Inc.",
-            job_title="Corporate Lawyer",
-            position_type=None,
-            person=cls.person_2,
-        )
-        cls.position_4 = PositionFactory.create(
-            date_granularity_start="%Y-%m-%d",
-            court=cls.court_2,
-            date_start=datetime.date(2020, 12, 14),
-            predecessor=cls.person_3,
-            appointer=cls.position_1,
-            judicial_committee_action="no_rep",
-            termination_reason="retire_mand",
-            position_type="c-jud",
-            person=cls.person_3,
-            how_selected="a_legis",
-            nomination_process="fed_senate",
-        )
-
-        cls.school_1 = SchoolFactory(name="New York Law School")
-        cls.school_2 = SchoolFactory(name="American University")
-
-        cls.education_1 = EducationFactory(
-            degree_level="jd",
-            person=cls.person_2,
-            degree_year=1965,
-            school=cls.school_1,
-        )
-        cls.education_2 = EducationFactory(
-            degree_level="ba",
-            person=cls.person_2,
-            school=cls.school_2,
-        )
-        cls.education_3 = EducationFactory(
-            degree_level="ba",
-            person=cls.person_3,
-            school=cls.school_1,
-        )
-
-        cls.political_affiliation_1 = PoliticalAffiliationFactory.create(
-            political_party="d",
-            source="b",
-            date_start=datetime.date(1993, 1, 1),
-            person=cls.person_1,
-            date_granularity_start="%Y",
-        )
-        cls.political_affiliation_2 = PoliticalAffiliationFactory.create(
-            political_party="d",
-            source="b",
-            date_start=datetime.date(2015, 12, 14),
-            person=cls.person_2,
-            date_granularity_start="%Y-%m-%d",
-        )
-        cls.political_affiliation_3 = PoliticalAffiliationFactory.create(
-            political_party="i",
-            source="b",
-            date_start=datetime.date(2015, 12, 14),
-            person=cls.person_3,
-            date_granularity_start="%Y-%m-%d",
-        )
-
-        cls.aba_rating_1 = ABARatingFactory(
-            rating="nq",
-            person=cls.person_2,
-            year_rated="2015",
-        )
-        super().setUpTestData()
 
 
 class SearchTestCase(TestCase):
